+++ R/construct_buffer.R
@@ -6,7 +6,7 @@
 #' @param shape String specifying plot shape. Either "circle" or "square"
 #' @param size Size of sample plot. Equals the radius for circles or the
 #' side-length for squares in mapunits
-#' @param return_sf If true, sf objects are returned.
+#' @param return_vec If true, vector objects are returned.
 #' @param verbose Print warning messages.
 #'
 #' @return
@@ -22,71 +22,10 @@
 #' @keywords internal
 #'
 #' @export
-construct_buffer <- function(coords, shape, size, return_sf, verbose) UseMethod("construct_buffer")
+construct_buffer <- function(coords, shape , size, return_vec = TRUE, verbose = TRUE) {
 
-#' @name construct_buffer
-#' @export
-construct_buffer.MULTIPOINT <- function(coords, shape, size,
-                                        return_sf = TRUE, verbose = TRUE) {
-
-    coords <- matrix(sf::st_coordinates(coords)[, 1:2], ncol = 2)
-
-    construct_buffer_internal(coords, shape, size,
-                              return_sf = return_sf, verbose = verbose)
-}
-
-#' @name construct_buffer
-#' @export
-construct_buffer.POINT <- function(coords, shape, size,
-                                   return_sf = TRUE, verbose = TRUE) {
-
-    coords <- matrix(sf::st_coordinates(coords)[, 1:2], ncol = 2)
-
-    construct_buffer_internal(coords, shape, size,
-                              return_sf = return_sf, verbose = verbose)
-}
-
-#' @name construct_buffer
-#' @export
-construct_buffer.sf <- function(coords, shape, size,
-                                return_sf = TRUE, verbose = TRUE) {
-
-    if (all(sf::st_geometry_type(coords) %in% c("POINT", "MULTIPOINT"))) {
-
-        coords <- matrix(sf::st_coordinates(coords)[, 1:2], ncol = 2)
-
-        construct_buffer_internal(coords, shape, size,
-                                  return_sf = return_sf, verbose = verbose)
-    }
-
-    else{stop("Only POINT or MULTIPOINT features supported!!11!!1!!")}
-}
-
-#' @name construct_buffer
-#' @export
-construct_buffer.sfc <- function(coords, shape, size,
-                                 return_sf = TRUE, verbose = TRUE) {
-
-    if (all(sf::st_geometry_type(coords) %in% c("POINT", "MULTIPOINT"))) {
-
-        coords <- matrix(sf::st_coordinates(coords)[, 1:2], ncol = 2)
-
-        construct_buffer_internal(coords, shape, size,
-                                  return_sf = return_sf, verbose = verbose)
-    }
-
-    else{stop("Only POINT or MULTIPOINT features supported!!11!!1!!")}
-}
-
-#' @name construct_buffer
-#' @export
-construct_buffer.matrix <- function(coords, shape , size, return_sf = TRUE, verbose = TRUE) {
-
-    construct_buffer_internal(coords, shape, size,
-                              return_sf = return_sf, verbose = verbose)
-}
-
-construct_buffer_internal <- function(coords, shape , size, return_sf = TRUE, verbose = TRUE) {
+    # make sure coords are matrix
+    coords <- points_as_mat(pts = coords)
 
     if (verbose) {
 
@@ -115,28 +54,25 @@ construct_buffer_internal <- function(coords, shape , size, return_sf = TRUE, ve
 
         sample_plots <- cbind(
             matrix(
-                c(coords[, 1] - size,
-                  coords[, 1] - size,
-                  coords[, 1] + size,
-                  coords[, 1] + size),
+                c(coords[, 1] - size, coords[, 1] - size,
+                  coords[, 1] + size, coords[, 1] + size),
                 ncol = 1),
 
             matrix(
-                c(coords[, 2] - size,
-                  coords[, 2] + size,
-                  coords[, 2] + size,
-                  coords[, 2] - size),
+                c(coords[, 2] - size, coords[, 2] + size,
+                  coords[, 2] + size, coords[, 2] - size),
                 ncol = 1),
             rep(seq_len(nrow(coords)), times = 4)
         )
     }
 
-    else{
+    else {
 
         stop(paste0("Shape option ", shape, " unkown."), call. = FALSE)
+
     }
 
-    if (return_sf) {
+    if (return_vec) {
 
         sample_plots <- split(sample_plots[, -3], sample_plots[, 3])
 
@@ -144,11 +80,11 @@ construct_buffer_internal <- function(coords, shape , size, return_sf = TRUE, ve
 
             mat <- matrix(x, ncol = 2)
 
-            sf::st_polygon(list(rbind(mat, mat[1, ])))
+            terra::vect(x = mat, type = "polygon")
 
         })
 
-        sample_plots <- sf::st_sf(data.frame(id = 1:nrow(coords), geom =  sf::st_sfc(sample_plots)))
+        sample_plots <- terra::vect(sample_plots)
 
     }