+++ tests/testthat/test-get-centroids.R
@@ -1,10 +1,12 @@
 context("get_centroids")
 
+centroids <- get_centroids(landscape, verbose = FALSE)
+
 test_that("get_centroids runs for all data types", {
 
-    raster_layer <- get_centroids(landscape)
-    raster_stack <- get_centroids(landscape_stack)
-    raster_list <- get_centroids(landscape_list)
+    raster_layer <- get_centroids(landscape, verbose = FALSE)
+    raster_stack <- get_centroids(landscape_stack, verbose = FALSE)
+    raster_list <- get_centroids(landscape_list, verbose = FALSE)
 
     expect_is(raster_layer, "tbl_df")
     expect_is(raster_stack, "tbl_df")
@@ -12,8 +14,6 @@ test_that("get_centroids runs for all data types", {
 })
 
 test_that("get_centroids returns in every column the correct type", {
-
-    centroids <- get_centroids(landscape)
 
     expect_type(centroids$layer, "integer")
     expect_type(centroids$level, "character")
@@ -24,8 +24,6 @@ test_that("get_centroids returns in every column the correct type", {
 })
 
 test_that("get_centroids returns centroid for each patch", {
-
-    centroids <- get_centroids(landscape)
 
     np <- lsm_l_np(landscape)
 
@@ -47,8 +45,8 @@ test_that("get_centroids allows to set cell_center", {
 
 test_that("get_centroids can return sf", {
 
-    centroids_sp <- get_centroids(landscape, cell_center = TRUE,
-                                  return_sf = TRUE)
+    centroids_spat <- get_centroids(landscape, cell_center = TRUE, return_vec = TRUE,
+                                  verbose = FALSE)
 
-    expect_is(centroids_sp, "sf")
+    expect_is(centroids_spat, "SpatVector")
 })