+++ NEWS.md
@@ -1,6 +1,8 @@
 # landscapemetrics 2.0.0
 * Improvements
     * `terra` and `sf` instead of `raster` and `sp` as underlying frameworks
+* New functions
+    * `points_as_mat()` helper function
 
 # landscapemetrics 1.5.6
 * Bugfixes